Drop commerce_order_ui_order_form support
commerce_order_ui_order_form is used by admins to edit current orders. This forms is highly buggy right now, and we don't have really need to it, as once order is completed it goes to ERP and can't be edited in Drupal; before that order should be normally handled by the customer, not by admins.